+++ src/forms/install.ts
@@ -15,4 +15,5 @@ export default function (app: any) {
     app.component('number-input', defineAsyncComponent(() => import('./number-input.vue')))
     app.component('date-picker', defineAsyncComponent(() => import('./date-picker/date-picker.vue')))
     app.component('time-input', defineAsyncComponent(() => import('./time-input.vue')))
+    app.component('search-input', defineAsyncComponent(() => import('./search-input.vue')))
 }