+++ src/interfaces/pagination.ts
@@ -3,6 +3,8 @@ import { _Sizes } from "./sizes"
 
 export interface _Pagination extends _Sizes {
     pager: _Pager
+    firstText?: string
     nextText?: string
     previousText?: string
+    lastText?: string
 }